Share holdings reconciliation domain policy
